1284-imperial.json
Added a discrete "on" command learned from my remote, allowing Smart IR to turn the unit on whenever any other commands are issued while the unit is off. This adds a third beep whenever adjustments are made to a unit that's already on, but I think it's well worth it to be able to turn the unit off completely and still be able to turn it on again. 

In this particular case, I converted the Celsius units to Fahrenheit for more personal use; they aren't perfect, but they're close enough for easy inference and control given the precision of a mini-split A/C unit.